// src/modules/notifications/notification-types.enum.ts
export enum NotificationType {
  // Auth
  WELCOME = 'WELCOME',
  BIRTHDAY = 'BIRTHDAY',

  // Events
  EVENT_CREATED = 'EVENT_CREATED',
  EVENT_MATCHED = 'EVENT_MATCHED',
  EVENT_JOIN_RESERVED = 'EVENT_JOIN_RESERVED',
  EVENT_JOIN_CONFIRMED = 'EVENT_JOIN_CONFIRMED',
  EVENT_CANCELLED = 'EVENT_CANCELLED',
  EVENT_REMINDER = 'EVENT_REMINDER',
  EVENT_RATE_REMINDER = 'EVENT_RATE_REMINDER',
  EVENT_CHAT_OPENED = 'EVENT_CHAT_OPENED',

  // Check-in
  CHECKIN_CONFIRMED = 'CHECKIN_CONFIRMED',
  CHECKIN_FAILED = 'CHECKIN_FAILED',

  // Payments
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
  PAYMENT_EXPIRED = 'PAYMENT_EXPIRED',

  // Blogs
  BLOG_PUBLISHED = 'BLOG_PUBLISHED',
  BLOG_REJECTED = 'BLOG_REJECTED',
  BLOG_COMMENTED = 'BLOG_COMMENTED',
  BLOG_NEW_POST_FOLLOWED_AUTHOR = 'BLOG_NEW_POST_FOLLOWED_AUTHOR',
  BLOG_NEW_POST_HOBBY_CATEGORY = 'BLOG_NEW_POST_HOBBY_CATEGORY',

  // Rewards
  REWARD_TIER_UPDATED = 'REWARD_TIER_UPDATED',

  // Security
  PASSWORD_CHANGED = 'PASSWORD_CHANGED',
  TWOFA_ENABLED = 'TWOFA_ENABLED',

  // Admin/Moderation
  CONTENT_APPROVED = 'CONTENT_APPROVED',
  CONTENT_REJECTED = 'CONTENT_REJECTED',
  CONTENT_TAKEN_DOWN = 'CONTENT_TAKEN_DOWN',
}

export enum NotificationCategory {
  MATCHED_EVENTS = 'matched_events',
  CREATED_EVENTS = 'created_events',
  OTHER = 'other',
}

// Notifications that should also trigger an email
export const EMAIL_NOTIFICATION_TYPES = [
  NotificationType.EVENT_CANCELLED,
  NotificationType.PAYMENT_EXPIRED,
  NotificationType.CONTENT_TAKEN_DOWN,
];